+++ vitest.config.ts
@@ -1,8 +1,11 @@
+import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
+  plugins: [react()],
   test: {
-    include: ['test/**/*.test.ts'],
+    include: ['test/**/*.test.ts', 'test/**/*.test.tsx'],
     environment: 'node',
+    environmentMatchGlobs: [['test/**/*.test.tsx', 'jsdom']],
   },
 });